refactor: реорганизация структуры проекта по смысловым пакетам
- src/main.py → src/altbooster.py
- src/system/ + src/backend.py + src/config.py → src/core/
- src/ui/*_page.py → src/tabs/*.py
- src/builtin_actions/ упразднён: BUILTIN_REGISTRY → ui/dynamic_page.py
- src/dynamic_page.py, src/widgets.py → src/ui/
- вкладка резервного копирования разбита на src/tabs/timesync/
- устаревшие SVG-иконки из корня icons/ удалены
