+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 Все значимые изменения в этом проекте документируются здесь.
 Формат основан на [Keep a Changelog](https://keepachangelog.com/ru/1.0.0/).
 
-## [5.6.8-beta] — 2026-03-15
+## [5.6.8] — 2026-03-17
 
 ### Добавлено
 - **Боковое меню навигации**: вкладки перенесены в боковую панель с иконкой и подписью; ширина панели регулируется и сохраняется между запусками; при сужении подписи скрываются автоматически (только иконки)
@@ -39,6 +39,14 @@
 - Корректное определение состояния иконок Papirus (установлен / применён / не применён) при первом запуске
 - Утечки файловых дескрипторов в `checks.py` и `rows.py`
 - `AttributeError` при сохранении пресета/экспорте
+
+### Внутреннее
+- **Рефакторинг структуры проекта**: файлы переорганизованы по смысловым пакетам:
+  - `src/main.py` → `src/altbooster.py` (точка входа)
+  - `src/ui/*_page.py` → `src/tabs/*.py` (без суффикса `_page`)
+  - `src/widgets.py` → `src/ui/widgets.py`, `src/dynamic_page.py` → `src/ui/dynamic_page.py`
+  - `src/config.py`, `src/backend.py` + системный слой → `src/core/` (переименован из `src/system/`)
+  - `src/builtin_actions/` упразднён: `BUILTIN_REGISTRY` перенесён в `src/ui/dynamic_page.py`, файлы — в `src/tabs/terminal_actions.py` и `src/tabs/amd_actions.py`
 
 ## [5.6.7] — 2026-03-08
