+++ README.md
@@ -90,31 +90,58 @@ altbooster
 
 ```
 altbooster/
-├── icons/                 # Графические ресурсы (иконки .svg, .png)
-├── src/                   # Исходный код приложения
-│   ├── builtin_actions/   # Функции-обработчики для действий из динамических вкладок
-│   ├── modules/           # JSON-файлы, описывающие UI для динамических страниц и задач
-│   ├── system/            # Бэкенд-модули для взаимодействия с системой
-│   │   ├── checks.py      # Функции для проверки состояния системы (установлен ли пакет и т.д.)
-│   │   ├── gsettings.py   # Обертка для работы с GSettings
-│   │   ├── packages.py    # Зарезервировано для будущего функционала по работе с пакетами
-│   │   ├── privileges.py  # Модуль для безопасного выполнения команд с sudo
-│   │   └── tweaks.py      # Реализация конкретных системных твиков
-│   ├── ui/                # Модули, отвечающие за отрисовку UI
-│   │   ├── dialogs.py     # Кастомные диалоговые окна (ввод пароля, редактор)
-│   │   ├── rows.py        # Библиотека переиспользуемых виджетов (TaskRow, AppRow, SettingRow)
-│   │   └── *.py           # Файлы, описывающие каждую статическую вкладку
-│   ├── backend.py         # Фасад, предоставляющий единый и чистый API для всего бэкенда
-│   ├── config.py          # Управление конфигурацией, путями и состоянием сессии
-│   ├── dynamic_page.py    # Движок, генерирующий UI на основе JSON-конфигураций
-│   ├── main.py            # Точка входа в приложение
-│   └── widgets.py         # Простые фабрики для создания стандартных виджетов Adwaita/GTK
-├── install.sh             # Скрипт установки
-├── uninstall.sh           # Скрипт удаления
-├── pyproject.toml         # Метаданные проекта и зависимости (PEP 621)
-├── CHANGELOG.md           # История изменений
-├── CONTRIBUTING.md        # Руководство для участников
-└── README.md              # Этот файл
+├── icons/                     # Графические ресурсы (иконки .svg, .png)
+│   ├── altbooster.svg/.png    # Иконка приложения
+│   └── hicolor/               # Иконки для системных тем GTK
+├── src/                       # Исходный код приложения
+│   ├── altbooster.py          # Точка входа в приложение
+│   ├── core/                  # Бэкенд-модули для взаимодействия с системой
+│   │   ├── backend.py         # Фасад, агрегирующий API всего бэкенда
+│   │   ├── borg.py            # Операции BorgBackup (низкоуровневые)
+│   │   ├── btrfs.py           # Снэпшоты и операции Btrfs
+│   │   ├── checks.py          # Функции проверки состояния системы
+│   │   ├── config.py          # Пути, версия, state.json, константы
+│   │   ├── gsettings.py       # Обёртки для gsettings/dconf
+│   │   ├── packages.py        # Логика epm и flatpak
+│   │   ├── privileges.py      # Безопасное выполнение команд через sudo/pkexec
+│   │   └── tweaks.py          # Реализация конкретных системных твиков
+│   ├── modules/               # JSON-файлы, описывающие UI для Data-Driven страниц
+│   ├── tabs/                  # Модули вкладок приложения
+│   │   ├── amd.py             # Вкладка «AMD»
+│   │   ├── amd_actions.py     # Действия для AMD-страницы
+│   │   ├── apps.py            # Вкладка «Приложения»
+│   │   ├── davinci.py         # Вкладка «DaVinci Resolve»
+│   │   ├── extensions.py      # Вкладка «Расширения GNOME»
+│   │   ├── flatpak.py         # Вкладка «Flatpak»
+│   │   ├── intel.py           # Вкладка «Intel» (scx_meteor)
+│   │   ├── maintenance.py     # Вкладка «Обслуживание»
+│   │   ├── setup.py           # Вкладка «Начало»
+│   │   ├── terminal.py        # Вкладка «Терминал»
+│   │   ├── terminal_actions.py# Действия для терминальной страницы
+│   │   └── timesync/          # Вкладка «Резервная копия» (Time Machine / BorgBackup)
+│   │       ├── page.py        # Главная страница вкладки
+│   │       ├── borg.py        # UI-логика резервного копирования
+│   │       ├── schedule.py    # Настройка расписания (systemd-таймеры)
+│   │       ├── snapshots.py   # Просмотр и управление снэпшотами
+│   │       ├── restore.py     # Восстановление из резервной копии
+│   │       ├── pickers.py     # Диалоги выбора файлов/директорий
+│   │       ├── metadata.py    # Метаданные архивов
+│   │       └── summary.py     # Сводный диалог после операций
+│   └── ui/                    # Переиспользуемые UI-компоненты
+│       ├── common.py          # Общие вспомогательные функции UI
+│       ├── dialogs.py         # Кастомные диалоговые окна (пароль, редактор)
+│       ├── dynamic_page.py    # Движок, генерирующий UI на основе JSON
+│       ├── install_preview_dialog.py # Диалог предпросмотра установки
+│       ├── rows.py            # Переиспользуемые Adw.ActionRow / Adw.ExpanderRow
+│       ├── widgets.py         # Фабрики стандартных виджетов Adwaita/GTK
+│       └── window.py          # AltBoosterWindow: layout, аутентификация, логи
+├── install.sh                 # Скрипт установки
+├── uninstall.sh               # Скрипт удаления
+├── Makefile                   # Сборка и установка через make
+├── pyproject.toml             # Метаданные проекта и зависимости (PEP 621)
+├── CHANGELOG.md               # История изменений
+├── CONTRIBUTING.md            # Руководство для участников
+└── README.md                  # Этот файл
 ```
 
 ## Лицензия
