+++ altbooster
@@ -1,3 +1,3 @@
 #!/bin/bash
-exec python3 /usr/share/altbooster/main.py "$@"
+exec python3 /usr/share/altbooster/altbooster.py "$@"
