+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = 'setuptools.build_meta'
 
 [project]
 name = "altbooster"
-version = "5.6.8-beta"
+version = "5.6.8"
 description = "Утилита настройки и обслуживания ALT Linux"
 readme = "README.md"
 license = "MIT"
@@ -21,7 +21,6 @@ include = ['altbooster*']
 
 [tool.setuptools.package-data]
 "altbooster.modules" = ['*.json']
-"altbooster.system" = ['*.tmp']
 
 [tool.ruff]
 line-length = 100
