+++ src/tabs/davinci.py
@@ -11,9 +11,9 @@ gi.require_version("Gtk", "4.0")
 gi.require_version("Adw", "1")
 from gi.repository import Adw, Gio, GLib, Gtk
 
-import backend
-import config
-from widgets import (
+from core import backend
+from core import config
+from ui.widgets import (
     make_icon, make_button, make_status_icon,
     set_status_ok, set_status_error, clear_status, make_suffix_box, make_scrolled_page,
 )
