+++ src/tabs/flatpak.py
@@ -12,11 +12,11 @@ gi.require_version("Gtk", "4.0")
 gi.require_version("Adw", "1")
 from gi.repository import Adw, GLib, Gtk
 
-import backend
-import config
+from core import backend
+from core import config
 from ui.common import load_module
 from ui.rows import TaskRow, SettingRow
-from widgets import make_icon, make_scrolled_page, make_suffix_box
+from ui.widgets import make_icon, make_scrolled_page, make_suffix_box
 
 
 @dataclass
