+++ src/tabs/timesync/__init__.py
@@ -1,0 +1,1 @@
+from .page import BorgPage
