+++ src/tabs/timesync/page.py
@@ -11,14 +11,16 @@ gi.require_version("Gtk", "4.0")
 gi.require_version("Adw", "1")
 from gi.repository import Adw, Gdk, GLib, Gtk, Pango
 
-import backend
-import config
-from widgets import (
+from core import backend
+from core import config
+from ui.widgets import (
     make_icon, make_scrolled_page, make_button,
     make_status_icon, set_status_ok, set_status_error, clear_status,
     make_suffix_box,
 )
-from ui.flatpak_page import _build_icon_index
+from .restore import BtrfsRestoreDialog, BorgArchiveBrowserDialog, BorgRestoreDialog
+from .pickers import HomeDirPickerDialog, FlatpakDataPickerDialog, FolderPickerDialog
+from .summary import BorgBackupSummaryDialog
 
 _INTERVALS = [
     (1,   "Каждый час"),
@@ -32,1378 +34,6 @@ _BTRFS_INTERVALS = [
     (6, "Каждые 6 часов"),
     (24, "Ежедневно"),
 ]
-
-
-class BtrfsRestoreDialog(Adw.AlertDialog):
-    def __init__(self, parent, snapshot: dict, log_fn):
-        super().__init__(
-            heading=f"Восстановить снимок",
-            body=f"Снимок: {snapshot.get('date_str', snapshot.get('name'))}",
-        )
-        self._snapshot = snapshot
-        self._log = log_fn
-        self._target_dir = str(Path.home())
-
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=6)
-        box.set_margin_top(4)
-
-        for icon_name, text in [
-            ("emblem-ok-symbolic",    "Все файлы ~/"),
-            ("emblem-ok-symbolic",    "Настройки GNOME и расширений"),
-            ("emblem-ok-symbolic",    "Данные и бинарники Flatpak (user)"),
-            ("dialog-error-symbolic", "Системные пакеты RPM — не затрагиваются"),
-            ("dialog-error-symbolic", "System-wide Flatpak — не затрагивается"),
-        ]:
-            row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-            row.append(make_icon(icon_name, 16))
-            lbl = Gtk.Label(label=text)
-            lbl.set_halign(Gtk.Align.START)
-            row.append(lbl)
-            box.append(row)
-
-        sep = Gtk.Separator()
-        sep.set_margin_top(6)
-        sep.set_margin_bottom(2)
-        box.append(sep)
-
-        warn = Gtk.Label(label="⚠ Файлы ~/  будут перезаписаны. После восстановления перезайдите в систему.")
-        warn.add_css_class("caption")
-        warn.add_css_class("dim-label")
-        warn.set_halign(Gtk.Align.START)
-        warn.set_wrap(True)
-        box.append(warn)
-
-        self.set_extra_child(box)
-        self.add_response("cancel", "Отмена")
-        self.add_response("extract", "Извлечь в папку…")
-        self.add_response("restore", "Восстановить ~/")
-        self.set_response_appearance("restore", Adw.ResponseAppearance.DESTRUCTIVE)
-        self.set_default_response("cancel")
-        self.connect("response", self._on_response)
-        self.set_transient_for(parent)
-
-    def _on_response(self, _d, response):
-        if response == "cancel":
-            return
-        if response == "extract":
-            self._pick_folder_then_restore()
-            return
-        self._do_restore(str(Path.home()))
-
-    def _pick_folder_then_restore(self):
-        try:
-            fd = Gtk.FileDialog()
-            fd.set_title("Папка для извлечения")
-            fd.select_folder(self.get_root(), None, self._on_folder_selected, None)
-        except AttributeError:
-            fc = Gtk.FileChooserNative(
-                title="Папка для извлечения",
-                action=Gtk.FileChooserAction.SELECT_FOLDER,
-                transient_for=self.get_root(),
-                accept_label="Выбрать",
-                cancel_label="Отмена",
-            )
-            def _resp(d, r):
-                if r == Gtk.ResponseType.ACCEPT:
-                    self._do_restore(d.get_file().get_path())
-                d.unref()
-            fc.connect("response", _resp)
-            fc.show()
-
-    def _on_folder_selected(self, dialog, result, _):
-        try:
-            folder = dialog.select_folder_finish(result)
-            if folder:
-                self._do_restore(folder.get_path())
-        except GLib.Error:
-            pass
-
-    def _do_restore(self, target_dir: str):
-        win = self.get_root()
-        if hasattr(win, "start_progress"):
-            win.start_progress("Восстановление снимка...")
-        self._log(f"\n▶  Восстановление снимка {self._snapshot['name']} → {target_dir}...\n")
-
-        def on_done(ok):
-            GLib.idle_add(self._finish, ok, win, target_dir == str(Path.home()))
-
-        backend.btrfs_snapshot_restore(self._snapshot["path"], target_dir, self._log, on_done)
-
-    def _finish(self, ok, win, is_home: bool):
-        if ok:
-            msg = "✔  Восстановление завершено!"
-            if is_home:
-                msg += " Перезайдите в систему для применения настроек GNOME."
-            self._log(msg + "\n")
-        else:
-            self._log("✘  Ошибка при восстановлении\n")
-        if hasattr(win, "stop_progress"):
-            win.stop_progress(ok)
-
-
-class BorgArchiveBrowserDialog(Adw.Window):
-
-    def __init__(self, parent, repo_path: str, archive_name: str):
-        super().__init__(transient_for=parent, modal=True)
-        self.set_title(f"Архив: {archive_name}")
-        self.set_default_size(700, 580)
-
-        self._repo_path = repo_path
-        self._archive_name = archive_name
-        self._all_items: list[dict] = []
-        self._children: dict = {}
-        self._nav_stack: list[str] = []
-        self._current_path: str = ""
-
-        self._back_btn = Gtk.Button()
-        self._back_btn.set_icon_name("go-previous-symbolic")
-        self._back_btn.connect("clicked", self._go_back)
-        self._back_btn.set_sensitive(False)
-
-        self._path_label = Gtk.Label(label="/")
-        self._path_label.set_ellipsize(Pango.EllipsizeMode.START)
-        self._path_label.set_hexpand(True)
-
-        self._search_btn = Gtk.ToggleButton()
-        self._search_btn.set_icon_name("system-search-symbolic")
-        self._search_btn.connect("toggled", self._toggle_search)
-
-        header = Adw.HeaderBar()
-        header.pack_start(self._back_btn)
-        header.set_title_widget(self._path_label)
-        header.pack_end(self._search_btn)
-
-        self._search_entry = Gtk.SearchEntry()
-        self._search_entry.set_hexpand(True)
-        self._search_entry.connect("search-changed", self._on_search)
-
-        self._search_bar = Gtk.SearchBar()
-        self._search_bar.set_child(self._search_entry)
-        self._search_bar.connect_entry(self._search_entry)
-        self._search_bar.set_search_mode(False)
-
-        self._list_box = Gtk.ListBox()
-        self._list_box.add_css_class("boxed-list")
-        self._list_box.set_selection_mode(Gtk.SelectionMode.NONE)
-
-        scroll = Gtk.ScrolledWindow()
-        scroll.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-        scroll.set_vexpand(True)
-        scroll.set_margin_start(12)
-        scroll.set_margin_end(12)
-        scroll.set_margin_bottom(12)
-        scroll.set_child(self._list_box)
-
-        self._spinner = Gtk.Spinner()
-        self._spinner.start()
-        self._spinner.set_valign(Gtk.Align.CENTER)
-        self._spinner.set_halign(Gtk.Align.CENTER)
-        self._spinner.set_vexpand(True)
-
-        self._stack = Gtk.Stack()
-        self._stack.add_named(self._spinner, "loading")
-        self._stack.add_named(scroll, "list")
-        self._stack.set_visible_child_name("loading")
-
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL)
-        box.append(header)
-        box.append(self._search_bar)
-        box.append(self._stack)
-        self.set_content(box)
-
-        threading.Thread(target=self._load, daemon=True).start()
-
-    def _load(self):
-        items = backend.borg_list_archive(self._repo_path, self._archive_name)
-        GLib.idle_add(self._on_loaded, items)
-
-    def _on_loaded(self, items: list[dict]):
-        self._all_items = items
-        self._children = self._build_tree(items)
-        self._dir_sizes = self._compute_dir_sizes(items)
-        self._navigate_to("")
-
-    def _compute_dir_sizes(self, items: list[dict]) -> dict:
-        sizes = {}
-        for item in items:
-            if item.get("type", "-") != "-":
-                continue
-            path = item.get("path", "").rstrip("/")
-            size = item.get("size") or 0
-            parts = path.split("/")
-            for i in range(len(parts)):
-                key = "/".join(parts[:i])
-                sizes[key] = sizes.get(key, 0) + size
-        return sizes
-
-    def _build_tree(self, items: list[dict]) -> dict:
-        children = {}
-
-        def _ensure_chain(path):
-            if path in children:
-                return
-            children[path] = {"dirs": [], "files": []}
-            if path:
-                parent = "/".join(path.split("/")[:-1])
-                _ensure_chain(parent)
-                if path not in children[parent]["dirs"]:
-                    children[parent]["dirs"].append(path)
-
-        for item in items:
-            path = item.get("path", "").rstrip("/")
-            if not path:
-                continue
-            itype = item.get("type", "-")
-            parent = "/".join(path.split("/")[:-1])
-            if itype == "d":
-                _ensure_chain(path)
-            else:
-                _ensure_chain(parent)
-                children[parent]["files"].append(item)
-
-        for v in children.values():
-            v["dirs"].sort()
-            v["files"].sort(key=lambda x: x.get("path", ""))
-
-        return children
-
-    def _navigate_to(self, path: str):
-        self._current_path = path
-        self._path_label.set_text("/" + path if path else "/")
-        self._back_btn.set_sensitive(bool(self._nav_stack))
-        self._populate_dir(path)
-        self._stack.set_visible_child_name("list")
-
-    def _clear_list(self):
-        while True:
-            row = self._list_box.get_row_at_index(0)
-            if row is None:
-                break
-            self._list_box.remove(row)
-
-    def _populate_dir(self, path: str):
-        self._clear_list()
-        data = self._children.get(path, {"dirs": [], "files": []})
-        dirs = data["dirs"]
-        files = data["files"]
-
-        if not dirs and not files:
-            row = Adw.ActionRow()
-            row.set_title("Папка пуста")
-            self._list_box.append(row)
-            return
-
-        for d in dirs:
-            name = d.split("/")[-1]
-            sub = self._children.get(d, {"dirs": [], "files": []})
-            count = len(sub["dirs"]) + len(sub["files"])
-            dir_size = self._dir_sizes.get(d, 0)
-            subtitle = f"{count} элем."
-            if dir_size:
-                subtitle += f"  ·  {_fmt_size(dir_size)}"
-            row = Adw.ActionRow()
-            row.set_title(name)
-            row.set_subtitle(subtitle)
-            row.set_activatable(True)
-            icon = make_icon("folder-symbolic", 16)
-            row.add_prefix(icon)
-            chevron = make_icon("go-next-symbolic", 16)
-            chevron.add_css_class("dim-label")
-            row.add_suffix(chevron)
-            row.connect("activated", lambda _, d=d: self._click_dir(d))
-            self._list_box.append(row)
-
-        for item in files:
-            path_full = item.get("path", "")
-            name = path_full.split("/")[-1]
-            size = item.get("size", 0)
-            mtime = (item.get("mtime") or "")[:16].replace("T", " ")
-            row = Adw.ActionRow()
-            row.set_title(name)
-            row.set_subtitle(mtime)
-            icon = make_icon(self._file_icon(name), 16)
-            row.add_prefix(icon)
-            lbl = Gtk.Label(label=_fmt_size(size))
-            lbl.add_css_class("dim-label")
-            lbl.add_css_class("caption")
-            lbl.set_valign(Gtk.Align.CENTER)
-            row.add_suffix(lbl)
-            self._list_box.append(row)
-
-    def _file_icon(self, name: str) -> str:
-        ext = name.rsplit(".", 1)[-1].lower() if "." in name else ""
-        if ext in ("jpg", "jpeg", "png", "gif", "webp", "svg", "bmp", "ico"):
-            return "image-x-generic-symbolic"
-        if ext in ("mp4", "mkv", "avi", "mov", "webm", "flv"):
-            return "video-x-generic-symbolic"
-        if ext in ("mp3", "flac", "ogg", "wav", "opus", "aac", "m4a"):
-            return "audio-x-generic-symbolic"
-        if ext in ("zip", "tar", "gz", "bz2", "xz", "7z", "rar", "zst"):
-            return "package-x-generic-symbolic"
-        if ext in ("py", "js", "ts", "c", "cpp", "h", "rs", "go", "java", "sh", "rb"):
-            return "text-x-script-symbolic"
-        if ext in ("pdf",):
-            return "x-office-document-symbolic"
-        return "text-x-generic-symbolic"
-
-    def _click_dir(self, path: str):
-        self._nav_stack.append(self._current_path)
-        self._navigate_to(path)
-
-    def _go_back(self, _=None):
-        if self._nav_stack:
-            self._navigate_to(self._nav_stack.pop())
-
-    def _toggle_search(self, btn):
-        active = btn.get_active()
-        self._search_bar.set_search_mode(active)
-        if active:
-            self._search_entry.grab_focus()
-        else:
-            self._search_entry.set_text("")
-            self._populate_dir(self._current_path)
-
-    def _on_search(self, entry):
-        q = entry.get_text().lower()
-        if not q:
-            self._populate_dir(self._current_path)
-            return
-        results = [i for i in self._all_items if q in i.get("path", "").lower()]
-        self._clear_list()
-        if not results:
-            row = Adw.ActionRow()
-            row.set_title("Ничего не найдено")
-            self._list_box.append(row)
-            return
-        for item in results[:300]:
-            path_full = item.get("path", "")
-            name = path_full.split("/")[-1]
-            size = item.get("size", 0)
-            mtime = (item.get("mtime") or "")[:16].replace("T", " ")
-            row = Adw.ActionRow()
-            row.set_title(name)
-            row.set_subtitle(path_full)
-            icon = make_icon(self._file_icon(name), 16)
-            row.add_prefix(icon)
-            lbl = Gtk.Label(label=f"{_fmt_size(size)}  {mtime}")
-            lbl.add_css_class("dim-label")
-            lbl.add_css_class("caption")
-            lbl.set_valign(Gtk.Align.CENTER)
-            row.add_suffix(lbl)
-            self._list_box.append(row)
-        if len(results) > 300:
-            row = Adw.ActionRow()
-            row.set_title(f"... и ещё {len(results) - 300}")
-            row.add_css_class("dim-label")
-            self._list_box.append(row)
-
-
-class BorgRestoreDialog(Adw.AlertDialog):
-
-    def __init__(self, repo_path: str, archive_name: str, log_fn, archive_date: str = ""):
-        super().__init__(
-            heading="Восстановить архив",
-            body="Файлы будут распакованы в указанную папку.",
-        )
-        self._repo_path = repo_path
-        self._archive_name = archive_name
-        self._log = log_fn
-
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=10)
-        box.set_margin_top(8)
-
-        info_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        icon = make_icon("drive-harddisk-symbolic", 16)
-        lbl = Gtk.Label(label=archive_date or archive_name)
-        lbl.add_css_class("dim-label")
-        info_row.append(icon)
-        info_row.append(lbl)
-        box.append(info_row)
-
-        sep = Gtk.Separator()
-        sep.set_margin_top(4)
-        box.append(sep)
-
-        self._cb_altbooster = Gtk.CheckButton(label="Настройки ALT Booster")
-        self._cb_altbooster.set_active(True)
-        self._cb_flatpak = Gtk.CheckButton(label="Данные Flatpak")
-        self._cb_flatpak.set_active(True)
-        self._cb_files = Gtk.CheckButton(label="Пользовательские файлы")
-        self._cb_files.set_active(True)
-        self._cb_packages = Gtk.CheckButton(label="Переустановить RPM-пакеты")
-        self._cb_packages.set_active(False)
-        self._cb_dconf = Gtk.CheckButton(label="Восстановить настройки GNOME (dconf)")
-        self._cb_dconf.set_active(True)
-        box.append(self._cb_altbooster)
-        box.append(self._cb_flatpak)
-        box.append(self._cb_files)
-        box.append(self._cb_packages)
-        box.append(self._cb_dconf)
-
-        sep2 = Gtk.Separator()
-        sep2.set_margin_top(4)
-        box.append(sep2)
-
-        target_label = Gtk.Label(label="Папка для восстановления:")
-        target_label.set_halign(Gtk.Align.START)
-        box.append(target_label)
-
-        path_row = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=6)
-        self._target_entry = Gtk.Entry()
-        self._target_entry.set_text(str(Path.home()))
-        self._target_entry.set_hexpand(True)
-        pick_btn = Gtk.Button(label="Выбрать…")
-        pick_btn.add_css_class("flat")
-        pick_btn.connect("clicked", self._on_pick_folder)
-        path_row.append(self._target_entry)
-        path_row.append(pick_btn)
-        box.append(path_row)
-
-        warn = Gtk.Label(label="⚠ Borg извлечёт файлы относительно выбранной папки")
-        warn.add_css_class("dim-label")
-        warn.add_css_class("caption")
-        warn.set_halign(Gtk.Align.START)
-        warn.set_wrap(True)
-        box.append(warn)
-
-        self.set_extra_child(box)
-        self.add_response("cancel", "Отмена")
-        self.add_response("restore", "Восстановить")
-        self.set_response_appearance("restore", Adw.ResponseAppearance.DESTRUCTIVE)
-        self.set_default_response("cancel")
-        self.connect("response", self._on_response)
-
-    def _on_pick_folder(self, _btn):
-        try:
-            fd = Gtk.FileDialog()
-            fd.set_title("Выберите папку для восстановления")
-            fd.select_folder(self.get_root(), None, self._on_folder_selected, None)
-        except AttributeError:
-            fc = Gtk.FileChooserNative(
-                title="Выберите папку для восстановления",
-                action=Gtk.FileChooserAction.SELECT_FOLDER,
-                transient_for=self.get_root(),
-                accept_label="Выбрать",
-                cancel_label="Отмена",
-            )
-            def _resp(d, r):
-                if r == Gtk.ResponseType.ACCEPT:
-                    self._target_entry.set_text(d.get_file().get_path())
-                d.unref()
-            fc.connect("response", _resp)
-            fc.show()
-
-    def _on_folder_selected(self, dialog, result, _):
-        try:
-            folder = dialog.select_folder_finish(result)
-            if folder:
-                self._target_entry.set_text(folder.get_path())
-        except GLib.Error:
-            pass
-
-    def _on_response(self, _d, response):
-        if response != "restore":
-            return
-        target_dir = self._target_entry.get_text().strip()
-        if not target_dir:
-            return
-        Path(target_dir).mkdir(parents=True, exist_ok=True)
-        win = self.get_root()
-        if hasattr(win, "start_progress"):
-            win.start_progress("Восстановление архива...")
-        self._log(f"\n▶  Восстановление {self._archive_name} → {target_dir}...\n")
-
-        restore_flatpak = self._cb_flatpak.get_active()
-        restore_packages = self._cb_packages.get_active()
-        restore_dconf = self._cb_dconf.get_active()
-
-        def _done(ok):
-            if not ok:
-                GLib.idle_add(self._finish, False, win)
-                return
-
-            meta_dir = Path(target_dir) / "tmp" / "altbooster-backup-meta"
-            if not meta_dir.exists():
-                meta_dir = Path(target_dir) / "altbooster-backup-meta"
-
-            def _step_dconf(ok_pkgs):
-                if restore_dconf and meta_dir.exists():
-                    self._log("▶  Восстановление настроек GNOME (dconf)...\n")
-                    ok_dconf = backend.restore_dconf_meta(meta_dir)
-                    if ok_dconf:
-                        self._log("   ✔ dconf загружен\n")
-                    else:
-                        self._log("   ⚠ Ошибка при загрузке dconf\n")
-                GLib.idle_add(self._finish, True, win)
-
-            def _step_packages(ok_flat):
-                if restore_packages and meta_dir.exists():
-                    backend.restore_packages_meta(meta_dir, self._log, _step_dconf)
-                else:
-                    _step_dconf(True)
-
-            if restore_flatpak and meta_dir.exists():
-                backend.restore_flatpak_meta(meta_dir, self._log, _step_packages)
-            else:
-                _step_packages(True)
-
-        backend.borg_extract(
-            self._repo_path, self._archive_name, target_dir, [],
-            self._log, _done,
-        )
-
-    def _finish(self, ok, win):
-        msg = "✔  Восстановление завершено!\n" if ok else "✘  Ошибка при восстановлении\n"
-        self._log(msg)
-        if hasattr(win, "stop_progress"):
-            win.stop_progress(ok)
-
-
-def _fmt_size(size: int) -> str:
-    if not isinstance(size, (int, float)) or size < 0:
-        return ""
-    for unit in ("Б", "КБ", "МБ", "ГБ", "ТБ"):
-        if size < 1024:
-            return f"{size:.0f} {unit}"
-        size /= 1024
-    return f"{size:.1f} ПБ"
-
-
-def _fmt_archive_date(name: str) -> str:
-    parts = name.split("-", 1)
-    if len(parts) == 2:
-        ts = parts[1].replace("T", " ").replace("-", ".").replace(".", " ", 2)
-        return ts
-    return name
-
-
-_XDG_HOME_DEFAULTS = [
-    "Documents", "Документы",
-    "Downloads", "Загрузки",
-    "Pictures", "Изображения",
-    "Music", "Музыка",
-    "Videos", "Видео",
-    "Desktop", "Рабочий стол",
-]
-
-_FOLDER_ICONS: dict[str, str] = {
-    "Documents": "folder-documents-symbolic",
-    "Документы": "folder-documents-symbolic",
-    "Downloads": "folder-download-symbolic",
-    "Загрузки": "folder-download-symbolic",
-    "Pictures": "folder-pictures-symbolic",
-    "Изображения": "folder-pictures-symbolic",
-    "Music": "folder-music-symbolic",
-    "Музыка": "folder-music-symbolic",
-    "Videos": "folder-videos-symbolic",
-    "Видео": "folder-videos-symbolic",
-    "Desktop": "user-desktop-symbolic",
-    "Рабочий стол": "user-desktop-symbolic",
-    "Templates": "folder-templates-symbolic",
-    "Шаблоны": "folder-templates-symbolic",
-    "Public": "folder-publicshare-symbolic",
-}
-
-_HOME_PICKER_CSS = """
-flowboxchild {
-    border-radius: 10px;
-    padding: 2px;
-}
-flowboxchild:hover {
-    background-color: alpha(currentColor, 0.06);
-}
-flowboxchild.checked {
-    background-color: alpha(@accent_bg_color, 0.18);
-    outline: 2px solid @accent_color;
-    outline-offset: -2px;
-}
-"""
-
-
-class HomeDirPickerDialog(Adw.Window):
-
-    def __init__(self, parent, dirs: list[str], selected: list[str], on_apply):
-        super().__init__(transient_for=parent, modal=True)
-        self.set_title("Папки домашней директории")
-        self.set_default_size(740, 520)
-        self._on_apply = on_apply
-        self._selected: set[str] = set(selected)
-        self._child_map: dict[str, Gtk.FlowBoxChild] = {}
-
-        css = Gtk.CssProvider()
-        css.load_from_string(_HOME_PICKER_CSS)
-        Gtk.StyleContext.add_provider_for_display(
-            Gdk.Display.get_default(), css,
-            Gtk.STYLE_PROVIDER_PRIORITY_APPLICATION,
-        )
-
-        header = Adw.HeaderBar()
-        btn_all = Gtk.Button(label="Выбрать всё")
-        btn_all.add_css_class("flat")
-        btn_all.connect("clicked", lambda _: self._set_all(True))
-        btn_none = Gtk.Button(label="Снять всё")
-        btn_none.add_css_class("flat")
-        btn_none.connect("clicked", lambda _: self._set_all(False))
-        header.pack_start(btn_none)
-        header.pack_end(btn_all)
-
-        self._flow = Gtk.FlowBox()
-        self._flow.set_selection_mode(Gtk.SelectionMode.NONE)
-        self._flow.set_activate_on_single_click(True)
-        self._flow.set_max_children_per_line(7)
-        self._flow.set_min_children_per_line(3)
-        self._flow.set_column_spacing(4)
-        self._flow.set_row_spacing(4)
-        self._flow.set_margin_start(16)
-        self._flow.set_margin_end(16)
-        self._flow.set_margin_top(16)
-        self._flow.set_margin_bottom(8)
-        self._flow.set_homogeneous(True)
-        self._flow.connect("child-activated", self._toggle)
-
-        for name in dirs:
-            child = self._make_child(name)
-            self._flow.append(child)
-            self._child_map[name] = child
-            if name in self._selected:
-                child.add_css_class("checked")
-
-        scroll = Gtk.ScrolledWindow()
-        scroll.set_vexpand(True)
-        scroll.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-        scroll.set_child(self._flow)
-
-        btn_apply = Gtk.Button(label="Применить")
-        btn_apply.add_css_class("suggested-action")
-        btn_apply.connect("clicked", self._apply)
-        action_bar = Gtk.ActionBar()
-        action_bar.pack_end(btn_apply)
-
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL)
-        box.append(header)
-        box.append(scroll)
-        box.append(action_bar)
-        self.set_content(box)
-
-    def _make_child(self, name: str) -> Gtk.FlowBoxChild:
-        icon_name = _FOLDER_ICONS.get(name, "folder-symbolic")
-
-        icon = Gtk.Image.new_from_icon_name(icon_name)
-        icon.set_pixel_size(52)
-        icon.set_margin_top(10)
-
-        check_img = Gtk.Image.new_from_icon_name("object-select-symbolic")
-        check_img.set_pixel_size(16)
-        check_img.set_halign(Gtk.Align.END)
-        check_img.set_valign(Gtk.Align.START)
-        check_img.set_margin_top(4)
-        check_img.set_margin_end(4)
-        check_img.add_css_class("accent")
-        check_img.set_visible(name in self._selected)
-
-        overlay = Gtk.Overlay()
-        overlay.set_child(icon)
-        overlay.add_overlay(check_img)
-
-        label = Gtk.Label(label=name)
-        label.set_max_width_chars(11)
-        label.set_wrap(True)
-        label.set_wrap_mode(Pango.WrapMode.WORD_CHAR)
-        label.set_justify(Gtk.Justification.CENTER)
-        label.set_margin_top(4)
-        label.set_margin_bottom(10)
-        label.add_css_class("caption")
-
-        vbox = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=0)
-        vbox.set_halign(Gtk.Align.CENTER)
-        vbox.set_size_request(90, -1)
-        vbox.append(overlay)
-        vbox.append(label)
-
-        child = Gtk.FlowBoxChild()
-        child.set_child(vbox)
-        child._check_img = check_img
-        return child
-
-    def _toggle(self, _flow, child):
-        name = next((n for n, c in self._child_map.items() if c is child), None)
-        if name is None:
-            return
-        if name in self._selected:
-            self._selected.discard(name)
-            child.remove_css_class("checked")
-            child._check_img.set_visible(False)
-        else:
-            self._selected.add(name)
-            child.add_css_class("checked")
-            child._check_img.set_visible(True)
-
-    def _set_all(self, state: bool):
-        for name, child in self._child_map.items():
-            if state:
-                self._selected.add(name)
-                child.add_css_class("checked")
-                child._check_img.set_visible(True)
-            else:
-                self._selected.discard(name)
-                child.remove_css_class("checked")
-                child._check_img.set_visible(False)
-
-    def _apply(self, _):
-        self._on_apply(sorted(self._selected))
-        self.close()
-
-
-class FlatpakDataPickerDialog(Adw.Window):
-
-    def __init__(self, parent, dirs: list[str], selected: list[str], on_apply, icons: dict | None = None):
-        super().__init__(transient_for=parent, modal=True)
-        self.set_title("Данные Flatpak — выбор приложений")
-        self.set_default_size(728, 600)
-        self._on_apply = on_apply
-        self._selected: set[str] = set(selected)
-        self._row_map: dict[str, tuple[Gtk.ListBoxRow, Gtk.CheckButton]] = {}
-        self._icons: dict = icons or {}
-
-        header = Adw.HeaderBar()
-        btn_all = Gtk.Button(label="Выбрать всё")
-        btn_all.add_css_class("flat")
-        btn_all.connect("clicked", lambda _: self._set_all(True))
-        btn_none = Gtk.Button(label="Снять всё")
-        btn_none.add_css_class("flat")
-        btn_none.connect("clicked", lambda _: self._set_all(False))
-        header.pack_start(btn_none)
-        header.pack_end(btn_all)
-
-        self._search = Gtk.SearchEntry()
-        self._search.set_hexpand(True)
-        self._search.connect("search-changed", self._on_search)
-
-        self._list = Gtk.ListBox()
-        self._list.add_css_class("boxed-list")
-        self._list.set_selection_mode(Gtk.SelectionMode.NONE)
-        self._list.set_filter_func(self._filter)
-        self._list.set_margin_start(12)
-        self._list.set_margin_end(12)
-        self._list.set_margin_top(8)
-        self._list.set_margin_bottom(8)
-
-        for name in dirs:
-            row, cb = self._make_row(name)
-            self._list.append(row)
-            self._row_map[name] = (row, cb)
-
-        search_bar = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL)
-        search_bar.set_margin_start(12)
-        search_bar.set_margin_end(12)
-        search_bar.set_margin_top(8)
-        search_bar.append(self._search)
-
-        scroll = Gtk.ScrolledWindow()
-        scroll.set_vexpand(True)
-        scroll.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-        scroll.set_child(self._list)
-
-        btn_apply = Gtk.Button(label="Применить")
-        btn_apply.add_css_class("suggested-action")
-        btn_apply.connect("clicked", self._apply)
-        action_bar = Gtk.ActionBar()
-        action_bar.pack_end(btn_apply)
-
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL)
-        box.append(header)
-        box.append(search_bar)
-        box.append(scroll)
-        box.append(action_bar)
-        self.set_content(box)
-
-    def _make_row(self, name: str) -> tuple[Gtk.ListBoxRow, Gtk.CheckButton]:
-        row = Adw.ActionRow()
-        parts = name.rsplit(".", 1)
-        row.set_title(parts[-1] if len(parts) > 1 else name)
-        row.set_subtitle(name)
-        row.set_subtitle_selectable(False)
-
-        icon_path = self._icons.get(name)
-        if icon_path:
-            try:
-                gi.require_version("GdkPixbuf", "2.0")
-                from gi.repository import GdkPixbuf
-                pixbuf = GdkPixbuf.Pixbuf.new_from_file_at_scale(icon_path, 32, 32, True)
-                texture = Gdk.Texture.new_for_pixbuf(pixbuf)
-                img = Gtk.Image()
-                img.set_from_paintable(texture)
-                img.set_size_request(32, 32)
-                row.add_prefix(img)
-            except Exception:
-                row.add_prefix(make_icon("application-x-executable-symbolic"))
-        else:
-            row.add_prefix(make_icon("application-x-executable-symbolic"))
-
-        cb = Gtk.CheckButton()
-        cb.set_active(name in self._selected)
-        cb.set_valign(Gtk.Align.CENTER)
-        cb.connect("toggled", self._on_toggled, name)
-        row.add_suffix(cb)
-        row.set_activatable_widget(cb)
-        row._app_id = name
-        return row, cb
-
-    def _on_toggled(self, cb, name: str):
-        if cb.get_active():
-            self._selected.add(name)
-        else:
-            self._selected.discard(name)
-
-    def _set_all(self, state: bool):
-        for name, (row, cb) in self._row_map.items():
-            if row.get_visible():
-                cb.set_active(state)
-
-    def _on_search(self, entry):
-        self._list.invalidate_filter()
-
-    def _filter(self, row) -> bool:
-        q = self._search.get_text().lower()
-        if not q:
-            return True
-        return q in row._app_id.lower()
-
-    def _apply(self, _):
-        self._on_apply(sorted(self._selected))
-        self.close()
-
-
-class FolderPickerDialog(Adw.Window):
-
-    def __init__(self, parent, title: str, dirs: list[str], selected: list[str], on_apply):
-        super().__init__(transient_for=parent, modal=True)
-        self.set_title(title)
-        self.set_default_size(1100, 800)
-        self._on_apply = on_apply
-        self._checks: dict[str, Gtk.CheckButton] = {}
-
-        header = Adw.HeaderBar()
-        btn_all = Gtk.Button(label="Выбрать всё")
-        btn_all.add_css_class("flat")
-        btn_all.connect("clicked", lambda _: [c.set_active(True) for c in self._checks.values()])
-        btn_none = Gtk.Button(label="Снять всё")
-        btn_none.add_css_class("flat")
-        btn_none.connect("clicked", lambda _: [c.set_active(False) for c in self._checks.values()])
-        header.pack_start(btn_none)
-        header.pack_end(btn_all)
-
-        flow = Gtk.FlowBox()
-        flow.set_max_children_per_line(4)
-        flow.set_min_children_per_line(2)
-        flow.set_selection_mode(Gtk.SelectionMode.NONE)
-        flow.set_column_spacing(0)
-        flow.set_row_spacing(0)
-        flow.set_margin_start(12)
-        flow.set_margin_end(12)
-        flow.set_margin_top(8)
-        flow.set_margin_bottom(8)
-        flow.set_homogeneous(True)
-
-        for name in sorted(dirs):
-            check = Gtk.CheckButton(label=name)
-            check.set_active(name in selected)
-            check.set_margin_top(4)
-            check.set_margin_bottom(4)
-            check.set_margin_start(8)
-            check.set_margin_end(8)
-            self._checks[name] = check
-            flow.append(check)
-
-        scroll = Gtk.ScrolledWindow()
-        scroll.set_vexpand(True)
-        scroll.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-        scroll.set_child(flow)
-
-        btn_apply = Gtk.Button(label="Применить")
-        btn_apply.add_css_class("suggested-action")
-        btn_apply.connect("clicked", self._apply)
-
-        action_bar = Gtk.ActionBar()
-        action_bar.pack_end(btn_apply)
-
-        box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL)
-        box.append(header)
-        box.append(scroll)
-        box.append(action_bar)
-        self.set_content(box)
-
-    def _apply(self, _):
-        selected = [name for name, check in self._checks.items() if check.get_active()]
-        self._on_apply(selected)
-        self.close()
-
-
-class BorgBackupSummaryDialog(Adw.Window):
-
-    def __init__(self, parent, repo_path: str, opts: dict, on_confirm):
-        super().__init__(transient_for=parent, modal=True)
-        self.set_title("Сводка резервной копии")
-        self.set_default_size(1100, 600)
-
-        self._repo_path = repo_path
-        self._opts = opts
-        self._on_confirm = on_confirm
-
-        header = Adw.HeaderBar()
-        header.set_show_end_title_buttons(False)
-
-        cancel_btn = Gtk.Button(label="Отмена")
-        cancel_btn.connect("clicked", lambda _: self.close())
-        header.pack_start(cancel_btn)
-
-        self._confirm_btn = Gtk.Button(label="Создать резервную копию")
-        self._confirm_btn.add_css_class("suggested-action")
-        self._confirm_btn.set_sensitive(False)
-        self._confirm_btn.connect("clicked", self._on_confirm_clicked)
-        header.pack_end(self._confirm_btn)
-
-        self._spinner = Gtk.Spinner()
-        self._spinner.start()
-        self._spinner.set_valign(Gtk.Align.CENTER)
-        self._spinner.set_halign(Gtk.Align.CENTER)
-        self._spinner.set_vexpand(True)
-
-        self._content_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=20)
-        self._content_box.set_margin_start(20)
-        self._content_box.set_margin_end(20)
-        self._content_box.set_margin_top(16)
-        self._content_box.set_margin_bottom(20)
-
-        scroll = Gtk.ScrolledWindow()
-        scroll.set_policy(Gtk.PolicyType.NEVER, Gtk.PolicyType.AUTOMATIC)
-        scroll.set_vexpand(True)
-        scroll.set_child(self._content_box)
-
-        self._stack = Gtk.Stack()
-        self._stack.add_named(self._spinner, "loading")
-        self._stack.add_named(scroll, "content")
-        self._stack.set_visible_child_name("loading")
-
-        root_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL)
-        root_box.append(header)
-        root_box.append(self._stack)
-        self.set_content(root_box)
-
-        threading.Thread(target=self._load_data, daemon=True).start()
-
-    @staticmethod
-    def _du(path: str) -> str:
-        return BorgBackupSummaryDialog._du_pair(path)[0]
-
-    @staticmethod
-    def _du_pair(path: str) -> tuple[str, int]:
-        try:
-            r = subprocess.run(
-                ["du", "-sk", path],
-                capture_output=True, text=True, encoding="utf-8", timeout=30,
-            )
-            if r.returncode == 0:
-                kb = int(r.stdout.split("\t")[0].strip())
-                if kb >= 1024 * 1024:
-                    return f"{kb / 1024 / 1024:.1f} ГБ", kb
-                if kb >= 1024:
-                    return f"{kb / 1024:.0f} МБ", kb
-                return f"{kb} КБ", kb
-        except Exception:
-            pass
-        return "", 0
-
-    @staticmethod
-    def _fmt_kb(kb: int) -> str:
-        if kb >= 1024 * 1024:
-            return f"{kb / 1024 / 1024:.1f} ГБ"
-        if kb >= 1024:
-            return f"{kb / 1024:.0f} МБ"
-        return f"{kb} КБ"
-
-    @staticmethod
-    def _du_children(parent: str) -> dict[str, str]:
-        try:
-            r = subprocess.run(
-                ["du", "-sh", "--max-depth=1", parent],
-                capture_output=True, text=True, encoding="utf-8", timeout=60,
-            )
-            result = {}
-            for line in r.stdout.splitlines():
-                parts = line.split("\t", 1)
-                if len(parts) == 2:
-                    size = parts[0].strip()
-                    name = Path(parts[1].strip()).name
-                    if name and name != Path(parent).name:
-                        result[name] = size
-            return result
-        except Exception:
-            return {}
-
-    @staticmethod
-    def _fmt_bytes(kb_str: str) -> str:
-        try:
-            kb = int(kb_str)
-            if kb >= 1024 * 1024:
-                return f"{kb / 1024 / 1024:.1f} ГБ"
-            if kb >= 1024:
-                return f"{kb / 1024:.0f} МБ"
-            return f"{kb} КБ"
-        except Exception:
-            return ""
-
-    def _load_data(self):
-        data = {}
-        home = Path.home()
-        total_kb = 0
-        var_app_str = str(home / ".var" / "app")
-
-        raw_paths = list(self._opts.get("paths", []))
-        paths_with_size = []
-        paths_kb = 0
-        for p in raw_paths:
-            s, kb = self._du_pair(p)
-            paths_with_size.append((p, s))
-            if not p.startswith(var_app_str):
-                paths_kb += kb
-                total_kb += kb
-        data["paths"] = paths_with_size
-        data["paths_kb"] = paths_kb
-
-        if self._opts.get("flatpak_apps"):
-            source_mode = self._opts.get("flatpak_apps_source", 0)
-            data["flatpak_apps_source"] = source_mode
-            try:
-                r = subprocess.run(
-                    ["flatpak", "list", "--app", "--columns=name,application,size"],
-                    capture_output=True, text=True, encoding="utf-8", timeout=15,
-                )
-                if r.returncode != 0:
-                    r = subprocess.run(
-                        ["flatpak", "list", "--app", "--columns=name,application"],
-                        capture_output=True, text=True, encoding="utf-8", timeout=15,
-                    )
-                installed = {}
-                for line in r.stdout.splitlines():
-                    if not line.strip():
-                        continue
-                    parts = [p.strip() for p in line.split("\t")]
-                    name = parts[0] if len(parts) > 0 else ""
-                    app_id = parts[1] if len(parts) > 1 else ""
-                    size = parts[2] if len(parts) > 2 else ""
-                    if app_id:
-                        installed[app_id] = (name or app_id, size)
-            except Exception:
-                installed = {}
-
-            if source_mode == 1:
-                try:
-                    raw = backend.flatpak_apps_from_booster_list()
-                    apps = []
-                    for name, app_id in raw:
-                        if app_id in installed:
-                            inst_name, size = installed[app_id]
-                            apps.append((inst_name or name, app_id, size))
-                    data["flatpak_apps"] = apps
-                except Exception:
-                    data["flatpak_apps"] = []
-            else:
-                data["flatpak_apps"] = [
-                    (name, app_id, size) for app_id, (name, size) in installed.items()
-                ]
-
-        if self._opts.get("flatpak_remotes"):
-            try:
-                r = subprocess.run(
-                    ["flatpak", "remotes", "--columns=name,url"],
-                    capture_output=True, text=True, encoding="utf-8", timeout=10,
-                )
-                remotes = []
-                for line in r.stdout.splitlines():
-                    parts = line.split("\t", 1)
-                    if len(parts) == 2:
-                        remotes.append((parts[0].strip(), parts[1].strip(), ""))
-                    elif parts[0].strip():
-                        remotes.append((parts[0].strip(), "", ""))
-                data["flatpak_remotes"] = remotes
-            except Exception:
-                data["flatpak_remotes"] = []
-
-        if self._opts.get("extensions"):
-            try:
-                r = subprocess.run(
-                    ["gnome-extensions", "list", "--enabled"],
-                    capture_output=True, text=True, encoding="utf-8", timeout=10,
-                )
-                data["extensions"] = [u.strip() for u in r.stdout.splitlines() if u.strip()]
-            except Exception:
-                data["extensions"] = []
-
-        if self._opts.get("flatpak_data"):
-            var_app = home / ".var" / "app"
-            sizes = self._du_children(str(var_app))
-            try:
-                all_dirs = sorted(p.name for p in var_app.iterdir() if p.is_dir()) if var_app.exists() else []
-            except Exception:
-                all_dirs = []
-            flt = self._opts.get("flatpak_data_filter")
-            dirs = [d for d in all_dirs if d in flt] if flt is not None else all_dirs
-            data["flatpak_data_dirs"] = [(d, sizes.get(d, "")) for d in dirs]
-            flatpak_total_str, flatpak_kb = self._du_pair(str(var_app))
-            data["flatpak_data_total"] = flatpak_total_str if flt is None else ""
-            total_kb += flatpak_kb
-
-        if self._opts.get("home_dirs"):
-            home_dirs = self._opts["home_dirs"]
-            home_pairs = [(d, *self._du_pair(str(home / d))) for d in home_dirs]
-            data["home_dirs"] = [(d, s) for d, s, _ in home_pairs]
-            home_kb = sum(kb for _, _, kb in home_pairs)
-            data["home_dirs_kb"] = home_kb
-            total_kb += home_kb
-
-        if self._opts.get("custom_paths"):
-            custom = self._opts["custom_paths"]
-            custom_pairs = [(p, *self._du_pair(p)) for p in custom]
-            data["custom_paths"] = [(p, s) for p, s, _ in custom_pairs]
-            total_kb += sum(kb for _, _, kb in custom_pairs)
-
-        try:
-            r = subprocess.run(
-                ["rpm", "-qa", "--queryformat", "%{NAME}\n"],
-                capture_output=True, text=True, encoding="utf-8", timeout=15,
-            )
-            data["packages_count"] = len([line for line in r.stdout.splitlines() if line.strip()])
-        except Exception:
-            data["packages_count"] = 0
-
-        data["total_kb"] = total_kb
-        GLib.idle_add(self._populate, data)
-
-    def _make_card(self, title: str, subtitle: str = "", icon: str = "", size: str = "", chip: str = "") -> Gtk.Box:
-        card = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=0)
-        card.add_css_class("card")
-        card.set_hexpand(True)
-
-        inner = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-        inner.set_margin_start(12)
-        inner.set_margin_end(12)
-        inner.set_margin_top(10)
-        inner.set_margin_bottom(10)
-
-        if icon:
-            ic = make_icon(icon, 16)
-            ic.set_valign(Gtk.Align.CENTER)
-            inner.append(ic)
-
-        text_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=2)
-        text_box.set_hexpand(True)
-
-        lbl = Gtk.Label(label=title)
-        lbl.set_halign(Gtk.Align.START)
-        lbl.set_ellipsize(Pango.EllipsizeMode.END)
-        lbl.set_max_width_chars(28)
-        text_box.append(lbl)
-
-        if subtitle:
-            sub = Gtk.Label(label=subtitle)
-            sub.add_css_class("dim-label")
-            sub.add_css_class("caption")
-            sub.set_halign(Gtk.Align.START)
-            sub.set_ellipsize(Pango.EllipsizeMode.END)
-            sub.set_max_width_chars(28)
-            text_box.append(sub)
-
-        inner.append(text_box)
-
-        right = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=4)
-        right.set_valign(Gtk.Align.CENTER)
-        right.set_halign(Gtk.Align.END)
-
-        if chip:
-            chip_lbl = Gtk.Label(label=chip)
-            chip_lbl.add_css_class("accent")
-            chip_lbl.add_css_class("caption")
-            chip_lbl.set_halign(Gtk.Align.END)
-            right.append(chip_lbl)
-
-        if size:
-            size_lbl = Gtk.Label(label=size)
-            size_lbl.add_css_class("dim-label")
-            size_lbl.add_css_class("caption")
-            size_lbl.set_halign(Gtk.Align.END)
-            right.append(size_lbl)
-
-        if chip or size:
-            inner.append(right)
-
-        card.append(inner)
-        return card
-
-    def _make_section(self, title: str, badge: str = "") -> tuple[Gtk.Expander, Gtk.FlowBox]:
-        label_box = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-
-        title_lbl = Gtk.Label(label=title)
-        title_lbl.add_css_class("heading")
-        title_lbl.set_halign(Gtk.Align.START)
-        label_box.append(title_lbl)
-
-        if badge:
-            badge_lbl = Gtk.Label(label=badge)
-            badge_lbl.add_css_class("dim-label")
-            badge_lbl.add_css_class("caption")
-            badge_lbl.set_valign(Gtk.Align.CENTER)
-            label_box.append(badge_lbl)
-
-        flow = Gtk.FlowBox()
-        flow.set_selection_mode(Gtk.SelectionMode.NONE)
-        flow.set_max_children_per_line(4)
-        flow.set_min_children_per_line(1)
-        flow.set_homogeneous(True)
-        flow.set_column_spacing(6)
-        flow.set_row_spacing(6)
-
-        flow_wrap = Gtk.Box(orientation=Gtk.Orientation.VERTICAL)
-        flow_wrap.set_margin_top(8)
-        flow_wrap.append(flow)
-
-        expander = Gtk.Expander()
-        expander.set_label_widget(label_box)
-        expander.set_expanded(True)
-        expander.set_child(flow_wrap)
-
-        return expander, flow
-
-    def _add_cards(self, flow: Gtk.FlowBox, items: list, icon: str, limit: int = 200):
-        for item in items[:limit]:
-            if isinstance(item, tuple):
-                title = item[0] or (item[1] if len(item) > 1 else "")
-                subtitle = item[1] if len(item) > 1 and item[0] else ""
-                size = item[2] if len(item) > 2 else ""
-            else:
-                title, subtitle, size = item, "", ""
-            flow.append(self._make_card(title, subtitle, icon, size))
-        if len(items) > limit:
-            flow.append(self._make_card(f"… и ещё {len(items) - limit}", icon="view-more-symbolic"))
-
-    def _populate(self, data: dict):
-        # 1. Хранилище
-        total_kb = data.get("total_kb", 0)
-        total_badge = f"Всего: {self._fmt_kb(total_kb)}" if total_kb else ""
-        repo_section, repo_flow = self._make_section("Хранилище", total_badge)
-        repo_card = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=0)
-        repo_card.add_css_class("card")
-        repo_card.set_hexpand(True)
-        repo_inner = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
-        repo_inner.set_margin_start(12)
-        repo_inner.set_margin_end(12)
-        repo_inner.set_margin_top(10)
-        repo_inner.set_margin_bottom(10)
-        repo_inner.append(make_icon("drive-harddisk-symbolic", 16))
-        repo_path_lbl = Gtk.Label(label=self._repo_path)
-        repo_path_lbl.set_halign(Gtk.Align.START)
-        repo_path_lbl.set_wrap(True)
-        repo_path_lbl.set_wrap_mode(Pango.WrapMode.CHAR)
-        repo_path_lbl.set_selectable(True)
-        repo_inner.append(repo_path_lbl)
-        repo_card.append(repo_inner)
-        repo_flow.append(repo_card)
-        self._content_box.append(repo_section)
-
-        # 2. Папки домашнего каталога (перед конфигами)
-        home_dirs = data.get("home_dirs")
-        if home_dirs:
-            home_kb = data.get("home_dirs_kb", 0)
-            badge = f"{len(home_dirs)} папок"
-            if home_kb:
-                badge += f"  ·  {self._fmt_kb(home_kb)}"
-            section, flow = self._make_section("Папки домашнего каталога", badge)
-            self._add_cards(flow, [(f"~/{d[0]}", "", d[1]) for d in home_dirs], "folder-home-symbolic")
-            self._content_box.append(section)
-
-        # 3. Конфигурационные файлы
-        var_app_str = str(Path.home() / ".var" / "app")
-        paths = [(p, s) for p, s in data.get("paths", [])
-                 if not p.startswith(var_app_str)]
-        if paths:
-            paths_kb = data.get("paths_kb", 0)
-            badge = f"{len(paths)} источников"
-            if paths_kb:
-                badge += f"  ·  {self._fmt_kb(paths_kb)}"
-            section, flow = self._make_section("Конфигурационные файлы", badge)
-            self._add_cards(flow, paths, "folder-symbolic")
-            self._content_box.append(section)
-
-        # 4. Flatpak — репозитории + приложения (объединено с данными)
-        remotes = data.get("flatpak_remotes")
-        if remotes is not None:
-            section, flow = self._make_section("Репозитории Flatpak", f"{len(remotes)} источников")
-            self._add_cards(flow, remotes, "network-server-symbolic")
-            self._content_box.append(section)
-
-        apps = data.get("flatpak_apps")
-        flatpak_dirs = data.get("flatpak_data_dirs")
-        if apps is not None:
-            data_map = {d: s for d, s in flatpak_dirs} if flatpak_dirs is not None else {}
-            with_data = sum(1 for _, app_id, _ in apps if app_id in data_map)
-            total = data.get("flatpak_data_total", "")
-
-            source_mode = data.get("flatpak_apps_source", 0)
-            source_label = "ALT Booster" if source_mode == 1 else "установленные"
-            badge = str(len(apps)) + f" приложений  ·  {source_label}"
-            if flatpak_dirs is not None:
-                badge += f"  ·  с данными: {with_data}"
-                if total:
-                    badge += f" ({total})"
-            else:
-                badge += "  ·  только список"
-
-            section, flow = self._make_section("Приложения Flatpak", badge)
-            for name, app_id, size in apps:
-                actual_id = app_id or name
-                if actual_id in data_map:
-                    data_size = data_map[actual_id]
-                    chip = f"+ данные{('  ' + data_size) if data_size else ''}"
-                    flow.append(self._make_card(name, app_id, "application-x-executable-symbolic", size, chip))
-                else:
-                    flow.append(self._make_card(name, app_id, "application-x-executable-symbolic", size))
-            if len(apps) > 200:
-                flow.append(self._make_card(f"… и ещё {len(apps) - 200}", icon="view-more-symbolic"))
-            self._content_box.append(section)
-        elif flatpak_dirs is not None:
-            total = data.get("flatpak_data_total", "")
-            badge = str(len(flatpak_dirs)) + " приложений" + (f"  ·  {total}" if total else "")
-            section, flow = self._make_section("Данные Flatpak (~/.var/app)", badge)
-            self._add_cards(flow, flatpak_dirs, "folder-symbolic")
-            self._content_box.append(section)
-
-        # 5. Расширения
-        exts = data.get("extensions")
-        if exts is not None:
-            section, flow = self._make_section("Расширения GNOME Shell", f"{len(exts)} включено")
-            self._add_cards(flow, exts, "application-x-addon-symbolic")
-            self._content_box.append(section)
-
-        # 6. Дополнительные пути
-        custom = data.get("custom_paths")
-        if custom:
-            section, flow = self._make_section("Дополнительные пути")
-            self._add_cards(flow, custom, "folder-symbolic")
-            self._content_box.append(section)
-
-        # 7. Системные пакеты
-        packages_count = data.get("packages_count", 0)
-        if packages_count > 0:
-            section, flow = self._make_section("Системные пакеты", f"{packages_count} пакетов")
-            flow.append(self._make_card(f"{packages_count} пакетов", "", "package-x-generic-symbolic"))
-            self._content_box.append(section)
-
-        self._stack.set_visible_child_name("content")
-        self._confirm_btn.set_sensitive(True)
-
-    def _on_confirm_clicked(self, _btn):
-        self.close()
-        self._on_confirm()
 
 
 class BorgPage(Gtk.Box):
@@ -1433,8 +63,8 @@ class BorgPage(Gtk.Box):
         self.append(clamp)
         self.append(self._stack)
 
-        tm_tab = self._build_time_machine_tab()
-        self._stack.add_titled_with_icon(tm_tab, "timemachine", "Time Machine", "document-revert-symbolic")
+        tm_tab = self._build_timesync_tab()
+        self._stack.add_titled_with_icon(tm_tab, "timesync", "TimeSync", "document-revert-symbolic")
 
         self._build_info_page()
         self._build_settings_page()
@@ -1444,7 +74,7 @@ class BorgPage(Gtk.Box):
             btrfs_tab, self._btrfs_body = self._build_btrfs_tab()
             self._stack.add_titled_with_icon(btrfs_tab, "btrfs", "Снимки", "camera-photo-symbolic")
 
-        self._page_tm = self._stack.get_page(self._stack.get_child_by_name("timemachine"))
+        self._page_tm = self._stack.get_page(self._stack.get_child_by_name("timesync"))
         self._page_info = self._stack.get_page(self._stack.get_child_by_name("info"))
         self._page_settings = self._stack.get_page(self._stack.get_child_by_name("settings"))
         self._page_schedule = self._stack.get_page(self._stack.get_child_by_name("schedule"))
@@ -1478,7 +108,7 @@ class BorgPage(Gtk.Box):
         self._build_prune_group()
         self._stack.add_titled_with_icon(scroll, "schedule", "График", "alarm-symbolic")
 
-    def _build_time_machine_tab(self) -> Gtk.Widget:
+    def _build_timesync_tab(self) -> Gtk.Widget:
         self._tm_box = Gtk.Box(orientation=Gtk.Orientation.VERTICAL, spacing=18)
         self._tm_box.set_margin_top(20)
         self._tm_box.set_margin_bottom(20)
@@ -1498,7 +128,7 @@ class BorgPage(Gtk.Box):
 
         # ── строка: заголовок + эксперт-кнопка ──────────────────────────
         top_bar = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=8)
-        page_title = Gtk.Label(label="Time Machine")
+        page_title = Gtk.Label(label="TimeSync")
         page_title.add_css_class("heading")
         page_title.set_halign(Gtk.Align.START)
         page_title.set_hexpand(True)
@@ -1992,7 +622,7 @@ class BorgPage(Gtk.Box):
             self._page_tm.set_visible(False)
         else:
             self._page_tm.set_visible(True)
-            self._stack.set_visible_child_name("timemachine")
+            self._stack.set_visible_child_name("timesync")
             self._tm_clamp.set_visible(False)
             for page in [self._page_info, self._page_settings, self._page_schedule]:
                 page.set_visible(False)
@@ -3045,7 +1675,7 @@ class BorgPage(Gtk.Box):
     def _save_repo_settings(self):
         config.state_set("borg_repo_path", self._row_repo_path.get_text().strip())
         config.state_set("borg_passphrase", self._row_passphrase.get_text())
-        if self._stack.get_visible_child_name() == "timemachine":
+        if self._stack.get_visible_child_name() == "timesync":
             GLib.idle_add(self._tm_refresh_archives)
 
     def _on_install_borg(self, _btn):
@@ -3308,6 +1938,7 @@ class BorgPage(Gtk.Box):
         icons_thread.start()
 
     def _load_flatpak_icons(self, all_dirs, selected):
+        from tabs.flatpak import _build_icon_index
         icons = _build_icon_index()
         GLib.idle_add(self._show_flatpak_data_dialog, all_dirs, selected, icons)
